+++ tests/Isolated/FeatureServiceProviderTest.php
@@ -6,13 +6,12 @@ use Illuminate\Support\Facades\Event;
 it('register listeners', function () {
     Event::fake();
     mockOnDemandDisk('features/TwoWords')->put('src/Listeners/Bar.php', '');
+    $provider = mockServiceProvider();
 
     $this->mock('alias:Illuminate\Foundation\Events\DiscoverEvents')
         ->shouldReceive('guessClassNamesUsing')->andReturn()
         ->shouldReceive('within')
         ->andReturn(['the-event' => [Listener::class]]);
-
-    $provider = mock(ServiceProvider::class, [mock()])->shouldAllowMockingProtectedMethods()->makePartial();
 
     $provider->bootListeners();
 
@@ -22,12 +21,4 @@ it('register listeners', function () {
 class Listener
 {
     public function handle(): void {}
-}
-
-class ServiceProvider extends FeatureServiceProvider
-{
-    protected function name(): string
-    {
-        return 'TwoWords';
-    }
 }
